Move DB functionality into separate module
