+++ lib/fathead/haddock/parse.hs
@@ -1,99 +1,13 @@
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 module Main where
 
-import Data.Char (ord, isAlpha)
-import Data.Csv
-import Data.String (IsString)
-import qualified Data.Text as DT
-import Data.Text (Text)
-import qualified Data.ByteString.Lazy.Char8 as BSZ
+import Data.Char (isAlpha)
 import Text.XML.HXT.Core
 import Data.Monoid ((<>))
 import Data.Tree.NTree.TypeDefs (NTree)
 import Network.URI (URI, parseURI)
 import Data.Maybe (fromJust)
 
-
-data EntryType = EntryArticle
-               | EntryDisambiguation
-               | EntryRedirect
-
-
--- | The type of entry in the DB.
-instance ToField EntryType where
-  toField EntryArticle = "A"
-  toField EntryDisambiguation = "D"
-  toField EntryRedirect = "R"
-
-
-type FieldText = String
-newtype Title = Title { getTitle :: String } deriving (ToField)
-type Disambugation = Text
-type Abstract = String
-
-
-newtype Categories = Categories { getCategories :: [Text] }
-
-
-instance ToField Categories where
-  toField (Categories { getCategories = cs }) = toField (DT.unlines cs)
-
-
-instance ToField URI where
-  toField = toField . show
-
-
-data Entry = Entry
-  { entryTitle :: !Title
-  , entryType  :: EntryType
-  , entryAlias :: Maybe Title
-  , entryCategories :: Maybe Categories
-  , entryDisambiguation :: Maybe Disambugation
-  , entryAbstract :: Maybe Abstract
-  , entryUrl :: Maybe URI
-  }
-
-
-emptyField :: Field
-emptyField = toField ("" :: FieldText)
-
-
-instance ToNamedRecord Entry where
-  toNamedRecord (Entry { entryTitle          = title
-                       , entryType           = eType
-                       , entryAlias          = alias
-                       , entryCategories     = categories
-                       , entryDisambiguation = disambiguation
-                       , entryAbstract       = abstract
-                       , entryUrl            = url }) =
-    namedRecord ([ "title"          .= toField title
-                 , "type"           .= toField eType
-                 , "redirect"       .= toField alias
-                 , "categories"     .= toField categories
-                 , "disambiguation" .= toField disambiguation
-                 , "abstract"       .= toField abstract
-                 , "source_url"     .= toField url
-                 ] ++ empties [ "null1", "null2", "null3", "external_links"
-                              , "see_also", "images" ])
-    where empties = fmap (.= emptyField)
-
-
-encodeOptions :: EncodeOptions
-encodeOptions = defaultEncodeOptions { encDelimiter = tab
-                                     , encIncludeHeader = True
-                                     , encUseCrLf = False
-                                     , encQuoting = QuoteNone
-                                     }
-  where tab = fromIntegral . ord $ '\t'
-
-
-outputHeader :: Header
-outputHeader = header [ "title" , "type", "redirect", "null1"
-                      , "categories", "null2", "see_also", "null3"
-                      , "external_links", "disambiguation"
-                      , "images", "abstract", "source_url"
-                      ]
+import FatHead.DB
 
 
 pagePath :: String -> FilePath
@@ -108,28 +22,6 @@ readHaddockDocument = readDocument sysConfig . pagePath
 
 hasClass :: ArrowXml a => String -> a XmlTree XmlTree
 hasClass c = hasAttrValue "class" (==c)
-
-
-article :: Title -> Abstract -> URI -> Entry
-article t a u = Entry { entryTitle =  t
-                    , entryType  = EntryArticle
-                    , entryAlias = Nothing
-                    , entryCategories = Nothing
-                    , entryDisambiguation = Nothing
-                    , entryAbstract = Just a
-                    , entryUrl = Just u
-                    }
-
-
-alias :: Title -> Title -> Entry
-alias orig new = Entry { entryTitle = new
-                       , entryType  = EntryRedirect
-                       , entryAlias = Just orig
-                       , entryCategories = Nothing
-                       , entryDisambiguation = Nothing
-                       , entryAbstract = Nothing
-                       , entryUrl = Nothing
-                       }
 
 
 buildAbstract :: ArrowXml a => a b XmlTree -> a b String
@@ -244,5 +136,4 @@ makeEntries = fmap concat . sequence $ entries
 
 
 main :: IO ()
-main = makeEntries >>= BSZ.writeFile "output.txt"
-       . encodeByNameWith encodeOptions outputHeader
+main = makeEntries >>= writeOutput
